extract devices into separate files
